backend/requirements.txt
fastapi
uvicorn
motor
pydantic
python-jose[cryptography]
passlib[bcrypt]
python-multipart
python-dotenv
email-validator
requests
Pillow
langdetect
google-genai==0.3.0
httpx>=0.27.0
